backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "espeasy-manager"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ampf>=0.1.21",
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "paho-mqtt>=2.1.0",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "unidecode>=1.3.8",
    "uvicorn>=0.34.0",
]

[[tool.uv.index]]
url = "https://europe-west3-python.pkg.dev/development-428212/pip/simple/"
